Add out of source build assertion
Problem
Running the cmake command inside the root project repository

Solution
Added a function that checks whether the project is being built inside the source directory, if that appears to be the case then a error is thrown back to the user.

Note
This only works for the root directory for now, other methods have to be investigated.